backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
numpy
scipy
pandas
scikit-learn
anthropic
requests
python-jose[cryptography]
passlib[bcrypt]
schedule
